+++ Broker/Position.cs
@@ -7,12 +7,12 @@ namespace TradingBot.Broker
     public class Position
     {
         public Contract Contract { get; set; }
-        public Decimal PositionAmount { get; set; }
-        public Decimal MarketPrice { get; set; }
-        public Decimal MarketValue { get; set; }
-        public Decimal AverageCost { get; set; }
-        public Decimal UnrealizedPNL { get; set; }
-        public Decimal RealizedPNL { get; set; }
+        public double PositionAmount { get; set; }
+        public double MarketPrice { get; set; }
+        public double MarketValue { get; set; }
+        public double AverageCost { get; set; }
+        public double UnrealizedPNL { get; set; }
+        public double RealizedPNL { get; set; }
 
         public override string ToString()
         {
